changed how deletion works, so that the draft is merely marked not deleted
this changes thinsg quite a bit in terms of how deletion works.  the aim is for this to allow deletes to be easily reverted, before they are published.

at this point unit tests pass, but more work is probably needed and more tests will need adding to catch erroneous behaviour.